Fixed HDRMetadata extraction; More async; Moved files around
