+++ AutoEncode/AutoEncodeServer/Models/EncodingJobModel.Build.cs
@@ -1,4 +1,5 @@
 ﻿using AutoEncodeServer.Models.Interfaces;
+using AutoEncodeServer.Utilities;
 using AutoEncodeServer.Utilities.Data;
 using AutoEncodeUtilities;
 using AutoEncodeUtilities.Base;
