+++ AutoEncode/AutoEncodeServer/Models/EncodingJobModel.Encode.cs
@@ -1,4 +1,5 @@
 ﻿using AutoEncodeServer.Models.Interfaces;
+using AutoEncodeServer.Utilities;
 using AutoEncodeUtilities;
 using AutoEncodeUtilities.Base;
 using AutoEncodeUtilities.Data;
